requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-openai
langchain-text-splitters
faiss-cpu
pypdf
openai
python-dotenv
llama-cpp-python
sentence-transformers
